feat: minio helper methods
